pyproject.toml
[tool.poetry]
name = "stocks-monitor-system"
version = "0.1.0"
description = ""
authors = ["Nazarov Roman <nazarovbusiness@gmail.com>"]
readme = "README.md"

# [tool.setuptools.packages.find]
# where = ["src"]

packages = [
    { include = "src", from = "." }
]

[tool.poetry.dependencies]
python = "^3.14.0"
fastapi = "^0.128.0"
uvicorn = "^0.40.0"
sqlalchemy = "^2.0.47"
psycopg2-binary = "^2.9.11"
pydantic-settings = "^2.13.1"
alembic = "^1.18.4"
pytest = "^9.0.2"
pytest-cov = "^7.0.0"
httpx = "^0.28.1"
pytest-asyncio = "^1.3.0"
python-dotenv = "^1.2.2"
httpx-socks = "^0.11.0"
factory-boy = "^3.3.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pwdlib = {extras = ["argon2"], version = "^0.3.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

